+++ BambooTracker/gui/configuration_dialog.cpp
@@ -436,6 +436,7 @@ void ConfigurationDialog::on_ConfigurationDialog_accepted()
 	pl->ptnUnmuteColor = ptnColors->child(25)->data(1, Qt::BackgroundRole).value<QColor>();
 	pl->ptnBackColor = ptnColors->child(26)->data(1, Qt::BackgroundRole).value<QColor>();
 	pl->ptnMarkerColor = ptnColors->child(27)->data(1, Qt::BackgroundRole).value<QColor>();
+	pl->ptnUnfocusedShadowColor = ptnColors->child(28)->data(1, Qt::BackgroundRole).value<QColor>();
 
 	QTreeWidgetItem* odrColors = ui->colorsTreeWidget->topLevelItem(1);
 	pl->odrDefTextColor = odrColors->child(0)->data(1, Qt::BackgroundRole).value<QColor>();
@@ -453,6 +454,7 @@ void ConfigurationDialog::on_ConfigurationDialog_accepted()
 	pl->odrBorderColor = odrColors->child(12)->data(1, Qt::BackgroundRole).value<QColor>();
 	pl->odrHeaderBorderColor = odrColors->child(13)->data(1, Qt::BackgroundRole).value<QColor>();
 	pl->odrBackColor = odrColors->child(14)->data(1, Qt::BackgroundRole).value<QColor>();
+	pl->odrUnfocusedShadowColor = odrColors->child(15)->data(1, Qt::BackgroundRole).value<QColor>();
 
 	QTreeWidgetItem* ilistColors = ui->colorsTreeWidget->topLevelItem(2);
 	pl->ilistTextColor = ilistColors->child(0)->data(1, Qt::BackgroundRole).value<QColor>();
@@ -694,6 +696,7 @@ void ConfigurationDialog::updateColorTree()
 	ptnColors->child(25)->setData(1, Qt::BackgroundRole, pl->ptnUnmuteColor);
 	ptnColors->child(26)->setData(1, Qt::BackgroundRole, pl->ptnBackColor);
 	ptnColors->child(27)->setData(1, Qt::BackgroundRole, pl->ptnMarkerColor);
+	ptnColors->child(28)->setData(1, Qt::BackgroundRole, pl->ptnUnfocusedShadowColor);
 
 	QTreeWidgetItem* odrColors = ui->colorsTreeWidget->topLevelItem(1);
 	odrColors->child(0)->setData(1, Qt::BackgroundRole, pl->odrDefTextColor);
@@ -711,6 +714,7 @@ void ConfigurationDialog::updateColorTree()
 	odrColors->child(12)->setData(1, Qt::BackgroundRole, pl->odrBorderColor);
 	odrColors->child(13)->setData(1, Qt::BackgroundRole, pl->odrHeaderBorderColor);
 	odrColors->child(14)->setData(1, Qt::BackgroundRole, pl->odrBackColor);
+	odrColors->child(15)->setData(1, Qt::BackgroundRole, pl->odrUnfocusedShadowColor);
 
 	QTreeWidgetItem* ilistColors = ui->colorsTreeWidget->topLevelItem(2);
 	ilistColors->child(0)->setData(1, Qt::BackgroundRole, pl->ilistTextColor);